backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
openai==2.33.0
pydantic==2.13.3
python-dotenv==1.2.2
langgraph==1.1.10
langchain-openai==1.2.1
langchain-core==1.3.2
langchain-community==0.4.1
langchain-tavily==0.2.18
